Add timestamp to the alarm class and database.
Adding a timestamp to indicate the last time the Alarm has been modified
in the database will allow correctly sync with the LightUpPi server.
This is to maintain the latest version of the alarms independently of
where they were modified (raspberry pi or android app) while the
connection between the two devices is broken.