Update database helper:
It now removes the tables and sets the default data for upgrades. This
is because they use case is minimal and thanks to the LightUpPi sync all
data can be easily restored.
It now also resets the Alarm and AlarmInstance tables, this replaces the
method to delete the database. This way all the other clock data is kept
and the alarm settings can reset to defaults from the menu (useful
during testing and debugging).